+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/listing/measure/LegacyMacraMeasureCriterionMapEntity.java
@@ -1,0 +1,54 @@
+package gov.healthit.chpl.listing.measure;
+
+import java.util.Date;
+
+import javax.persistence.Basic;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.OneToOne;
+import javax.persistence.Table;
+
+import org.hibernate.annotations.Immutable;
+
+import lombok.Data;
+
+@Entity
+@Immutable
+@Data
+@Table(name = "allowed_measure_criteria_legacy_map")
+public class LegacyMacraMeasureCriterionMapEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
+    private Long id;
+
+    @Basic(optional = false)
+    @Column(name = "allowed_criteria_measure_id", nullable = false)
+    private Long measureCriterionId;
+
+    @Basic(optional = true)
+    @OneToOne(optional = true, fetch = FetchType.LAZY)
+    @JoinColumn(name = "allowed_criteria_measure_id", unique = true, nullable = true, insertable = false, updatable = false)
+    private MeasureCriterionMapEntity measureCriterionMap;
+
+    @Column(name = "macra_criteria_map_id", nullable = false, updatable = false, insertable = false)
+    private Long legacyMacraMeasureId;
+
+    @Column(name = "creation_date", nullable = false, updatable = false, insertable = false)
+    private Date creationDate;
+
+    @Column(nullable = false)
+    private Boolean deleted;
+
+    @Column(name = "last_modified_date", nullable = false, updatable = false, insertable = false)
+    private Date lastModifiedDate;
+
+    @Column(name = "last_modified_user", nullable = false)
+    private Long lastModifiedUser;
+}