+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/upload/certifiedProduct/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Upload handlers to parse XML files for a variety of listing formats.
- */
-package gov.healthit.chpl.upload.certifiedProduct;