+++ RELEASE_NOTES.md
@@ -1,5 +1,18 @@
 # Release Notes
 
+## Version 26.0.0
+_14 December 2020_
+
+### Breaking Change
+* Move G1/G2 macra measures to measures at listing level
+
+### Features
+* Generate direct review download file and manage DR cache
+* Enable ROLE_ONC_STAFF user type
+* Make "username" not required for new users
+
+---
+
 ## Version 25.17.0
 _30 November 2020_